requirements.txt
typer
python-dotenv
requests
rich
google-genai 
openai